Limit upload speed of firmware image
The BR62 has problems to process incoming packets fast enough.
To mitigate this, it sends a lot of tcp window updates to indicate that
the receive window is full.
However this somehow interacts with the upload of the firmware image
which then fails on many occasions. As a temporary workaround we limit
the upload speed to a safe value such that the BR62 is able to handle
all incoming packets.